Initial commit for PuRRR project (#9)
"Pull Request Review Reminder" (PuRRR) - see README for an overview of
this new sample project.

Some TODOs mentioned in the code - not implemented yet to keep the size
of this PR manageable.

Emphasis in this project was on fault tolerance, readability and
reusability. In other words, this is more of a "real product" demo than
a toy/proof-of-concept.