ENG-1028 Samples fixes (#109)
This PR addresses a couple of bugs and includes a suggestion to clarify
the usage of the slash command.

I initially expected an argument with the slash command (I'm not very
familiar with using them). While we might choose to ignore users who are
unfamiliar with slash commands, I thought it worth mentioning for
consideration.